src/app/components/style-guide/colors/colors.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component, DebugElement } from '@angular/core';
import { By } from '@angular/platform-browser';
import { ColorsComponent } from './colors.component';

describe('Component: StyleGuideColors', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [ColorsComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder): void {
    builder = tcb;
  }));

  it('should inject the component', inject([ColorsComponent], (component: ColorsComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(ColorsTestControllerComponent)
      .then((fixture: ComponentFixture<any>) => {
        let query: DebugElement = fixture.debugElement.query(By.directive(ColorsComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  directives: [ColorsComponent],
  selector: 'td-test',
  template: `
    <td-style-guide-colors></td-style-guide-colors>
  `,
})
class ColorsTestControllerComponent {
}
